Check delete if user

+++ includes/objects/entry.php
@@ -222,14 +222,14 @@ class entry {
     }
 
     public function delete () {
-        $query = "DELETE FROM ".DB_PREFIX."company_has_data WHERE data_id = ?";
+        $query = "DELETE FROM ".DB_PREFIX."company_has_data INNER JOIN ".DB_PREFIX."data ON ".DB_PREFIX."data.id = data_id WHERE data_id = ? AND user_id = ?";
         if ($stmt = $this-> conn -> prepare($query)) {
-            $stmt->bind_param("i", $this->id);
+            $stmt->bind_param("ii", $this->id, $_SESSION['userId']);
             $stmt->execute();
         }
-        $query = "DELETE FROM ".DB_PREFIX."datatype_has_data WHERE data_id = ?";
+        $query = "DELETE FROM ".DB_PREFIX."datatype_has_data INNER JOIN ".DB_PREFIX."data ON ".DB_PREFIX."data.id = data_id WHERE data_id = ? AND user_id = ?";
         if ($stmt = $this-> conn -> prepare($query)) {
-            $stmt -> bind_param("i", $this->id);
+            $stmt -> bind_param("ii", $this->id, $_SESSION['userId']);
             $stmt -> execute();
         }
         $query = "DELETE FROM ".DB_PREFIX."data WHERE id = ? AND user_id = ?";
